requirements.txt
groq==0.12.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain_google_genai==2.0.4
langchain_text_splitters==0.3.2
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.40.1
faiss-cpu>=1.7.4